+++ src/app/migrations/telemetry-settings.migrations.ts
@@ -1,5 +1,6 @@
 import { cloneDeep } from 'lodash';
 import { TELEMETRY_SETTINGS_DEFAULT, TelemetrySettings } from '../models/telemetry-settings';
+import { AUTOMATION_CONFIGS_DEFAULT } from '../models/automations';
 
 const migrations: { [v: number]: (data: any) => any } = {
   1: toLatest,
@@ -7,6 +8,11 @@ const migrations: { [v: number]: (data: any) => any } = {
 
 export function migrateTelemetrySettings(data: any): TelemetrySettings {
   let currentVersion = data.version || 0;
+  // Reset to latest when the current version is higher than the latest
+  if (currentVersion > TELEMETRY_SETTINGS_DEFAULT.version) {
+    data = toLatest(data);
+    console.log(`Reset future telemetry settings version back to version ${currentVersion + ''}`);
+  }
   while (currentVersion < TELEMETRY_SETTINGS_DEFAULT.version) {
     data = migrations[++currentVersion](data);
     currentVersion = data.version;